Separate different layers
Dense and Conv2D were defined inside NNLayer but
they are now separated and given different source
and header files
